Refactor code and finish majority functionality

+++ week-1-ts-ecosystem/w01e05-my-own-cli/src/common.ts
@@ -1,14 +1,17 @@
+import {createShopCard} from "./shopCardFactory";
+
 const prompts = require(("prompts"))
 import { faker } from '@faker-js/faker';
 import {Key, ObjectData, ProductType} from "./type";
-import {writeFileSync, unlinkSync} from  "fs"
-
+import {writeFileSync,readFileSync, unlinkSync} from  "fs"
+import {join} from "path"
 import chalk from "chalk";
 import figlet, { Options} from "figlet";
 import {
-    amountFakeDataQuestion, deleteQuestion,
+    amountFakeDataQuestion, choiceQuestion,
     generateQuestion,
     initializeData,
+    jsonDataQuestion,
     LOOP_LENGTH,
     TypeQuestions,
 } from "./data";
@@ -19,11 +22,21 @@ export const createJSONFile  = async (data:ObjectData) => {
     const JSONResponse = JSON.stringify(data)
         try {
             await writeFileSync(path, JSONResponse)
-            return textMessage(`Data written successfully to ${file}`, "blue")
+             textMessage(`Data written successfully to ${file}`, "blue")
+            return "SUCCESS"
+
         } catch (err) {
-            return textMessage('An error has occurred ')
+             textMessage('An error has occurred ')
+            return "ERROR"
         }
 }
+
+export const readJSONFile = async () => {
+    const filePath = join(__dirname,"../persistent-data/cart-items.json")
+    const jsonData =  readFileSync(filePath).toString()
+        return JSON.parse(jsonData)
+}
+
 
 export const deleteDataFromFile = async () => {
     try {
@@ -48,16 +61,31 @@ export const textMessage = (text:string, blue?:string) => {
     )
 }
 
+export  const pinoMessage = (mesage: string) => {
+    const pino = require('pino')
+    const logger = pino({
+        transport: {
+            target: 'pino-pretty'
+        },
+        options: {
+            colorize: true
+        }
+    })
+    logger.info(mesage)
+}
+
 export const onSubmit = (prompt: any ,answer: string) => {
+    if (answer === "leave") return null
     if (prompt.type === 'text') textMessage(`Nice to meet you ${answer}`, "blue")
 }
 
 const createRandomUser = (Type: Key): ProductType => {
     return {
         id:  faker.datatype.uuid(),
+        type: Type,
         name: faker.commerce.productName(),
         amount: Number(faker.random.numeric()),
-        price: Type === "forFree" ? null : faker.commerce.price(),
+        price: Type ===  Key.GIVE_FOR_FREE ? 0 : Number(faker.commerce.price()),
 
     };
 }
@@ -66,8 +94,9 @@ const generateMyRecord = async (Type: Key,userName: string) => {
     let prepareCartData: ObjectData = initializeData
     const {questionsII} = generateQuestion(userName)
     const { productName, productAmount, productPrice} =  await prompts(questionsII)
-    const newProduct = {
+    const newProduct: ProductType = {
         id: faker.datatype.uuid(),
+        type: Type,
         name: productName,
         amount: productAmount,
         price: productPrice,
@@ -81,36 +110,51 @@ export const giveMeFakeData = (Type: Key, amount: number) => {
     Array.from({ length: amount }).forEach(() => {
         shopCartData[Type].push(createRandomUser(Type));
     });
+    console.log("shopCartData",shopCartData)
     return shopCartData
 }
 
+
 export  const generateMyData = async (userName:string)=> {
     for (let i = 0; i <= LOOP_LENGTH; i++) {
-        const {generateFakeData,Type} = await prompts(TypeQuestions)
+        const {productType} = await prompts(TypeQuestions[0])
+        if(productType === "leave") return null
+        const {generateFakeData} = await prompts(TypeQuestions[1])
         if( generateFakeData) {
             const {amountRecords} = await prompts(amountFakeDataQuestion)
-            if(!amountRecords) return
-            return  giveMeFakeData(Type,amountRecords)
+            const shopCartData = giveMeFakeData(productType,amountRecords)
+            return {shopCartData,productType}
         }
-        const response = await generateMyRecord(Type, userName)
-        if(response[Type as Key].length) {
+        const shopCartData = await generateMyRecord(productType, userName)
+        if(shopCartData[productType as Key].length) {
             textMessage(`You have been created your Record !`, "blue")
+            return {shopCartData,productType}
         }
+    }
+        textMessage((`See you Soon ${userName}`),"blue")
+        return null
+}
+
+export const beforeJsonFile = async (shopCartData : ObjectData) => {
+    const {choice} = await prompts(choiceQuestion)
+    if(choice === "leave") return null
+    if(choice === "json")  {
+      return  await createJSONFile(shopCartData)
     }
 }
 
+export const afterJsonFIle = async (shopCartData : ObjectData,Type: Key, Name:string) => {
+    const {jsonDataChoice} = await prompts(jsonDataQuestion)
+        if(jsonDataChoice === "delete") await deleteDataFromFile()
+        if(jsonDataChoice === "add")  {
+        const cartItemFromJsonFile: ObjectData = await readJSONFile()
+        await createShopCard(cartItemFromJsonFile,Type, Name)
+    }
+}
 export const Avada_Kedavra = async (Name: string) => {
-    const {loopQuestion} = generateQuestion(Name)
-    const prepareCartData = await generateMyData(Name)
-    if(!prepareCartData) return textMessage((`See you Soon ${Name}`),"blue")
-
-    await createJSONFile(prepareCartData)
-
-    const {loopAnswer} = await prompts(loopQuestion)
-    if(!loopAnswer) return textMessage((`See you Soon ${Name}`),"blue")
-    const {deleteData} = await prompts(deleteQuestion)
-        if(deleteData) {
-            await deleteDataFromFile()
-        }
+    const response = await generateMyData(Name)
+    if(!response)  return null
+    const {shopCartData, productType} = response
+    await beforeJsonFile(shopCartData) && await afterJsonFIle(shopCartData,productType,Name)
     await Avada_Kedavra(Name)
 }
